requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic-settings
psycopg2-binary
sqlalchemy
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart==0.0.9
numpy==1.26.4
prefect==2.19.5

# Testing libraries
pytest
pytest-asyncio
respx
httpx

# New additions
langchain
langchain-community
sentence-transformers
faiss-cpu
pypdf
reportlab
langchain-openai
python-dotenv
sumy==0.11.0
nltk
langchain-huggingface

# Data Science additions (DS1.1.1)
pandas
openpyxl
scikit-learn 